requirements.txt
streamlit
transformers
deep-translator
openai-whisper
torch
ffmpeg-python